Rename and Extract methods

+++ src/main/java/com/everest/airline/controller/FlightController.java
@@ -1,25 +1,27 @@
 package com.everest.airline.controller;
 
 
-import com.everest.airline.Data;
+import com.everest.airline.DataHandler;
 import com.everest.airline.model.Flight;
+import com.everest.airline.service.FlightService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 import java.io.File;
+import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.util.ArrayList;
 
 @RestController
 public class FlightController {
     @Autowired
-    Data data;
+    DataHandler data;
+    @Autowired
+    FlightService flightService;
 
-    String filePath = "/Users/Pravallika/Documents/Assignments/airlines/src/main/java/com/everest/airline/flights";
+    String filePath = "src/main/java/com/everest/airline/flights";
     File directory = new File(filePath);
+    File[] files = directory.listFiles();
 
     @GetMapping("/flights")
     public ArrayList<Flight> getAllFlights() throws IOException {
@@ -31,10 +33,15 @@ public class FlightController {
         return data.getDataFromFile(number);
     }
 
+    @PostMapping("/flights")
+    public long addFlight(String source, String destination, String departureDate, String departTime, String arrivalTime, int availableSeats, int firstClassSeats, int secondClassSeats, int economicClassSeats, double firstClassBasePrice, double secondClassBasePrice, double economicClassBasePrice) throws IOException {
+        return flightService.getNextFlightNumber();
+    }
+
     @DeleteMapping("/flights/{number}")
-    public boolean deleteFlight(@PathVariable("number") long number) {
+    public boolean deleteFlight(@PathVariable("number") long number) throws FileNotFoundException {
         File[] files = directory.listFiles((File pathname) -> pathname.getName().equals(String.valueOf(number)));
-        assert files != null;
+        if (files == null) throw new FileNotFoundException("No such file");
         File file = new File(files[0].getPath());
         return file.delete();
     }
